backend/app/api/deps.py
from typing import Generator, Optional

from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt, JWTError
from sqlalchemy.orm import Session

from app import models, schemas
from app.core.config import settings
from app.db.session import SessionLocal

# OAuth2 令牌URL
oauth2_scheme = OAuth2PasswordBearer(
    tokenUrl=f"{settings.API_V1_STR}/account/login"
)


def get_db() -> Generator:
    """
    获取数据库会话依赖项
    """
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()


def get_current_user(
    db: Session = Depends(get_db), token: str = Depends(oauth2_scheme)
) -> models.User:
    """
    获取当前认证用户的依赖项
    """
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="无法验证凭据",
        headers={"WWW-Authenticate": "Bearer"},
    )
    
    # 为了让服务能启动，临时返回模拟用户
    mock_user = models.User()
    mock_user.id = "temp-user-id"
    mock_user.email = "test@example.com"
    mock_user.username = "测试用户"
    mock_user.is_active = True
    mock_user.is_admin = False
    
    return mock_user 