backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.0
email-validator>=2.0.0
python-dotenv>=1.0.0
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
ruff==0.0.284
black==23.7.0
mypy==1.5.1 